Move data_311 to external/nyc311
